resole threads from task-2
